refactor: Rename utils
